requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.1
Django==2.1.3
django-cors-headers==2.1.0
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
mysqlclient==1.3.13
Pillow==5.3.0
pycparser==2.19
PyJWT==1.6.4
pytz==2018.7
requests==2.20.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.24.1
